+++ mkutils.lua
@@ -344,7 +344,7 @@ Make.testlogfile = testlogfile
 --env.Make:add("htlatex", "${htlatex} ${latex_par} '\\\makeatletter\\def\\HCode{\\futurelet\\HCode\\HChar}\\def\\HChar{\\ifx\"\\HCode\\def\\HCode\"##1\"{\\Link##1}\\expandafter\\HCode\\else\\expandafter\\Link\\fi}\\def\\Link#1.a.b.c.{\\g@addto@macro\\@documentclasshook{\\RequirePackage[#1,html]{tex4ht}\\let\\HCode\\documentstyle\\def\\documentstyle{\\let\\documentstyle\\HCode\\expandafter\\def\\csname tex4ht\\endcsname{#1,html}\\def\\HCode####1{\\documentstyle[tex4ht,}\\@ifnextchar[{\\HCode}{\\documentstyle[tex4ht]}}}\\makeatother\\HCode '${config}${tex4ht_sty_par}'.a.b.c.\\input ' ${input}")
 
 -- template for calling LaTeX with tex4ht loaded
-Make.latex_command = "${htlatex} ${latex_par} '\\makeatletter"..
+Make.latex_command = "${htlatex} --interaction=${interaction} ${latex_par} '\\makeatletter"..
 "\\def\\HCode{\\futurelet\\HCode\\HChar}\\def\\HChar{\\ifx\"\\HCode"..
 "\\def\\HCode\"##1\"{\\Link##1}\\expandafter\\HCode\\else"..
 "\\expandafter\\Link\\fi}\\def\\Link#1.a.b.c.{\\g@addto@macro"..
@@ -356,10 +356,13 @@ Make.latex_command = "${htlatex} ${latex_par} '\\makeatletter"..
 "\\input \"\\detokenize{${tex_file}}\"'"
 
 env.Make:add("htlatex",function(par)
+  local settings = get_filter_settings "htlatex" or {}
   local command = Make.latex_command
   if os.type == "windows" then
     command = command:gsub("'",'')
   end
+  -- the interaction parameter is configurable in settings or as a parameter
+  par.interaction = par.interaction or settings.interaction or "batchmode"
   command = command % par
   print("LaTeX call: "..command)
   os.execute(command)